test: remove rln-same test and rename rln-diff to rln
