+++ test/configs.ts
@@ -1,5 +1,3 @@
-import * as path from "path";
-
 // MERKLE TREE
 export const MERKLE_TREE_DEPTH = 20;
 export const MERKLE_TREE_ZERO_VALUE = BigInt(0);
